requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
tqdm==4.66.1

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
boto3==1.34.30
minio==7.2.2

# Data Processing
pandas==2.1.3
numpy==1.24.3
pyarrow==11.0.0
apache-beam==2.52.0
pyspark==3.5.1

# Streaming
confluent-kafka==2.3.0
kafka-python==2.0.2

# Orchestration & Workflow
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-apache-kafka==1.2.0
prefect==2.16.3

# Analytics & Transformation
dbt-core==1.7.3
dbt-postgres==1.6.7
dbt-databricks==1.7.3

# Azure & Databricks
azure-storage-file-datalake==12.15.0
azure-identity==1.15.0
databricks-sdk==0.18.0
databricks-sql-connector==3.0.2
delta-spark==3.0.0

# Visualization
dash==2.14.1
dash-bootstrap-components==1.5.0
plotly==5.18.0
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0

# Data Quality & Testing
great-expectations==0.18.8
faker==23.1.0
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
pylint==3.0.3